Added seeLine and useLine, complete checkLine
The method checkLine, that checks if a line has been completed after a
player played, has been completed and now checks every direction.

The methods seeLine and useLine have been added to the Square class. The
first one states if a square is already used to complete a line in a
certain direction, the other states if it now completes a line in a
certain direction.

A strange bug in the gameTurn method has not been fixed : according to
the compiler, there's an invisible "=" token.